requirements.txt
langchain==0.0.292
PyPDF2==3.0.1
streamlit==1.26.0
faiss-cpu==1.7.4
google-generativeai==0.1.0
dill==0.3.7
python-magic